[#52] Split scalafmt in its own subproject
