+++ build.sbt
@@ -5,7 +5,7 @@ lazy val root = Project(id = "kantan-sbt", base = file("."))
     publishLocal    := {},
     publishArtifact := false
   )
-  .aggregate(core, kantan, release, scalastyle)
+  .aggregate(core, kantan, release, scalafmt, scalastyle)
 
 lazy val core = project
   .settings(
@@ -17,7 +17,6 @@ lazy val core = project
   .settings(
     addSbtPlugin("com.eed3si9n"        % "sbt-unidoc"      % Versions.sbtUnidoc),
     addSbtPlugin("com.github.tkawachi" % "sbt-doctest"     % Versions.sbtDoctest),
-    addSbtPlugin("com.geirsson"        % "sbt-scalafmt"    % Versions.sbtScalafmt),
     addSbtPlugin("com.typesafe.sbt"    % "sbt-ghpages"     % Versions.sbtGhPages),
     addSbtPlugin("com.typesafe.sbt"    % "sbt-site"        % Versions.sbtSite),
     addSbtPlugin("de.heikoseeberger"   % "sbt-header"      % Versions.sbtHeader),
@@ -51,6 +50,18 @@ lazy val scalastyle = project
   )
   .dependsOn(core)
 
+lazy val scalafmt = project
+  .settings(
+    moduleName := "kantan.sbt-scalafmt",
+    name       := "scalafmt",
+    sbtPlugin  := true
+  )
+  .enablePlugins(AutomateHeaderPlugin)
+  .settings(
+    addSbtPlugin("com.geirsson" % "sbt-scalafmt" % Versions.sbtScalafmt)
+  )
+  .dependsOn(core)
+
 lazy val kantan = project
   .settings(
     moduleName := "kantan.sbt-kantan",
@@ -62,7 +73,7 @@ lazy val kantan = project
     addSbtPlugin("org.xerial.sbt" % "sbt-sonatype" % Versions.sbtSonatype),
     addSbtPlugin("com.jsuereth"   % "sbt-pgp"      % Versions.sbtPgp)
   )
-  .dependsOn(core, release, scalastyle)
+  .dependsOn(core, release, scalafmt, scalastyle)
 
 addCommandAlias(
   "validate",
