--- core/src/sbt-test/scalafmt/all/expected-plugins.txt
@@ -1,1 +1,0 @@
-addSbtPlugin("com.nrinaudo" % "kantan.sbt" % sys.props("plugin.version"))
