--- core/src/sbt-test/scalafmt/all/test
@@ -1,2 +1,0 @@
-> scalafmtAll
-> check
