--- core/src/sbt-test/scalafmt/compile/test
@@ -1,1 +1,0 @@
--> checkStyle
