--- core/src/sbt-test/scalafmt/ok/test
@@ -1,2 +1,0 @@
-> checkStyle
-> test:checkStyle
