--- core/src/sbt-test/scalafmt/sbt/build.sbt
@@ -1,1 +1,0 @@
-startYear      := Some(1901)
