--- core/src/sbt-test/scalafmt/sbt/test
@@ -1,1 +1,0 @@
--> checkStyle
