--- core/src/sbt-test/scalafmt/test/.scalafmt.conf
@@ -1,22 +1,0 @@
-style                           = defaultWithAlign
-maxColumn                       = 121
-includeCurlyBraceInSelectChains = false
-
-binPack {
-  parentConstructors = true
-}
-
-rewrite {
-  rules = [RedundantBraces, RedundantParens, SortImports, PreferCurlyFors]
-}
-
-rewriteTokens {
-  "->": "→"
-  "=>": "⇒"
-  "<-": "←"
-}
-
-align {
-  arrowEnumeratorGenerator = true
-  tokens.add               = [":="]
-}
