--- core/src/sbt-test/scalafmt/test/test
@@ -1,1 +1,0 @@
--> test:checkStyle
