+++ release/src/sbt-test/release/checkstyle-test/project/plugins.sbt
@@ -1,2 +1,5 @@
 addSbtPlugin(
   "com.nrinaudo" % "kantan.sbt-release" % sys.props("plugin.version"))
+
+addSbtPlugin(
+  "com.nrinaudo" % "kantan.sbt-scalafmt" % sys.props("plugin.version"))
