+++ scalafmt/src/sbt-test/scalafmt/main/src/main/scala/BadlyFormed.scala
