+++ scalafmt/src/sbt-test/scalafmt/ok/src/main/scala/WellFormed.scala
