+++ scalafmt/src/sbt-test/scalafmt/ok/src/test/scala/NotBadlyFormed.scala
