+++ scalafmt/src/sbt-test/scalafmt/sbt/project/plugins.sbt
@@ -1,0 +1,1 @@
+addSbtPlugin("com.nrinaudo" % "kantan.sbt-scalafmt" % sys.props("plugin.version"))
