+++ scalafmt/src/sbt-test/scalafmt/scalafmtAll/expected/Versions.txt
@@ -1,0 +1,3 @@
+object Versions {
+  val fooBar = "1.0.0"
+}
