+++ scalafmt/src/sbt-test/scalafmt/scalafmtAll/expected/bar.txt
