+++ scalafmt/src/sbt-test/scalafmt/scalafmtAll/expected/foo.txt
