+++ scalafmt/src/sbt-test/scalafmt/scalafmtAll/project/plugins.sbt
@@ -1,5 +1,2 @@
 addSbtPlugin(
-  "com.nrinaudo" % "kantan.sbt-release" % sys.props("plugin.version"))
-
-addSbtPlugin(
   "com.nrinaudo" % "kantan.sbt-scalafmt" % sys.props("plugin.version"))
