+++ scalafmt/src/sbt-test/scalafmt/scalafmtAll/src/main/scala/Foo.scala
