+++ scalafmt/src/sbt-test/scalafmt/scalafmtAll/src/test/scala/Bar.scala
