+++ scalafmt/src/sbt-test/scalafmt/test/src/test/scala/BadlyFormed.scala
