+++ scalastyle/src/sbt-test/scalastyle/main/project/plugins.sbt
@@ -1,2 +1,1 @@
-addSbtPlugin(
-  "com.nrinaudo" % "kantan.sbt-scalastyle" % sys.props("plugin.version"))
+addSbtPlugin("com.nrinaudo" % "kantan.sbt-scalastyle" % sys.props("plugin.version"))
